Add exercise 2
In this exercise, I learned about the idea that React.createElement is
just creating a sort of idea of an object rather than something that can
be directly plopped on a page. You need to use the React Render method
in order to actually get an object to appear on the page.

I'm not entirely clear what can go in to the props of the createElement
function - but I think it's ok to leave it as is for now.