Add exercise 5
I learned how to combine together variables and static definitions in a
React object - including how to use the spread property together with
said definitions. It's a bit strange with all of the nested { - but
something like "style" is an object {} which is then inside {} because
that allows it to be interpreted as JSX. A classname, on the other hand,
 is just a string - so it doesn't need to have {} inside of {} (although
  you can do {`foo ${bar}`} to add vars to the string.

One thing I found confusing but eventually figured out was the idea that
having text inside of a React object like <Box> Some Text </Box> means
that the Box object receives a parameter of "children" that has "Some
Text" in it. If you don't specify the children property and, instead,
just say something like ...foo at the end, you can then add a final
property in the div like <div ... {...foo}/> which will add the
children.

I think a better way to do that would be to actually specify the
children property and put it like <div> {children} </div> instead as
that is much less magical and easy to understand where stuff is coming
from.