Add exercise 6 + update git ignored
Updates gitignore to not include .swp files which are temp files on
Windows for saving state that I don't want to see.

I learned that the default behaviour of submitting a form is to make a
GET request with all the fields of the form as query params. I never
even thought about what might be happening there.

I learned there was a console.dir function and that it will output a
JSON blob instead of an HTML object for DOM elements. It also generally
provides more info than a generic console.log will.

I learned the purpose of "htmlFor" in React - which ties a label to an
input, allows you to focus the input when clicking on the label, and
helps screenreaders work correctly. I'm totally blown away by this and
have never heard of this before!

I learned that you can retrieve the value of fields in a form using the
specified ID via event.target.elements.<id>.value rather than relying on
any particular element in an array or by order or anything like that.

I learned the importance of making sure to recalculate the state on
input changes. I originally said if(!isValid) { setError(...) } and that
made sense in my mind cause I'm like oh if there is an error then set
it. I missed the fact that I never set it to empty if there is no
error, though.